Rust 1.75.0 is the current latest release of the Rust language and toolchains. This change is necessary since I have discovered that with the previously used version, 1.67.0, future releases will no longer be possible. The reason is that `cargo publish`, even when using the `--locked` flag, seems to perform its' own package version resolution. During this process it hits an issue with one of our dependencies that has bumped its' MSRV to 1.70.0.

While I think technically this is a bug in the Rust tooling, it's still a nice idea to do the upgrade. The commits in this PR perform the upgrade in "three" steps with Rust versions 1.69, 1.70 and 1.75. I have picked these versions as I think that each of these mentioned something in their `CHANGELOG` that offers a potential benefit or risk to zellij. For more information, refer to the individual commits.

The quick upshot is: When going from 1.67.0 to 1.75.0, build time (in Debug) drops from ~240 s to ~190 s on my PC (for a clean build), and initial package resolution (from an empty cargo cache) goes down from ~75 s to ~10 s. There are other benefits (e.g. cross-crate inlining) that promise performance improvements, but I do not know how to measure performance at this point. One particular feature that may be of interest is the introduction of [native WASM exceptions][1] in Rust 1.72.0 for our plugins, but I have not looked into this any futher so far.

I compiled zellij with each of the new toolchains and made sure they all execute with current `main`, running on an x86_64-unknown-linux host. Finally, I performed a simulated release with the version compiled with 1.75.0. I can upload the bins if there is interest.


[1]: https://github.com/rust-lang/rust/pull/111322/